EXP: Do not blindly honor the UNINIT flag
The super block is backupep into some block groups, but the bitmap of
these groups are marked as UNINIT.

If we ignore the flag and trust the containt of the group bitmap, we collect
all the used blocks reported by the file system. The md5sum of the restored
file system also match the one of the original one.