Improve progress computation
Avoid a "divide by zero" situation by storing the block count instead of
storing fraction of percentage. Another issue with the fractions was that
the float "rounding error" could cause the percentage to go above 100%.

start is converted to an unsigned long long to not do arithmetic with
disparate types (signed vs unsigned).

During progression computation, if the block count is zero, we assume we
are almost done (99.99%), otherwise compute the percentage by dividing
the current value by the total to reach.

Remove duplicated code while at it.