Fix buffer overflow issues in progress formatting
The main issue was that snprintf() was called with a NULL value for the
buffer to determine the length required for the string, but the returned
length was not validated to be smaller than the capacity of the buffer.
To avoid this, the value to be formatted is limited to 999, which will
produce the string " > 999 hrs ", which requires 12 bytes. I guess a
related issue was handled in the past by increasing the size of string in
the struct to 13 chars. The value 999 prevent the string to be truncated.

A secondary issue was that the code was a little bit duplicated.
format_duration() is used to that care of this, with 2 helpers functions
to take care of the details.